# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Vehicle API",
        "version": "1.0"
    },
    "paths": {
        "/api/vehicles/{id}/rent": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Rent a vehicle",
                "operationId": "1694c182d73686a0ccd31b25bf045c46",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Vehicle--id"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Id of User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "comment",
                        "in": "query",
                        "description": "Trip comment",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Grocery shopping trip"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Trip"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    },
                    "404": {
                        "description": "Vehicle not found"
                    }
                }
            }
        },
        "/api/vehicles/{id}/release": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Release a vehicle",
                "operationId": "1ac7d439e20178b69fac275d64075e4f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Vehicle--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Trip"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    },
                    "404": {
                        "description": "Vehicle not found"
                    }
                }
            }
        },
...
        "/api/trips/{id}": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Get Trip by id",
                "operationId": "6b8785740e1e7723d8899ea7bd468b48",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Trip--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Trip"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trip not found"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of User",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Users"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Insert a new User",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "description": "User to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get User by id",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/User--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Updates a User",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/User--id"
                    }
                ],
                "requestBody": {
                    "description": "User to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a User",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/User--id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Trip": {
                "required": [],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "comment": {
                        "type": "string",
                        "example": "Grocery shopping trip"
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "vehicle_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "start_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true,
                        "example": "2022-01-02 12:13:14"
                    },
                    "finish_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true,
                        "example": "2022-01-02 13:00:00",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true,
                        "example": "2022-01-02 12:13:14"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true,
                        "example": "2022-01-02 12:13:14"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Ivan Ivanov"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "email@email.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true,
                        "example": "2022-01-02 12:13:14"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true,
                        "example": "2022-01-02 12:13:14"
                    }
                },
                "type": "object"
            },
            "Users": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "Trip--id": {
                "name": "trip_id",
                "in": "path",
                "description": "Id of Trip",
                "required": true,
                "schema": {
                    "type": "integer"
                },
                "example": 1
            },
            "User--id": {
                "name": "user_id",
                "in": "path",
                "description": "Id of User",
                "required": true,
                "schema": {
                    "type": "integer"
                },
                "example": 1
            },
            "Vehicle--id": {
                "name": "vehicle_id",
                "in": "path",
                "description": "Id of Vehicle",
                "required": true,
                "schema": {
                    "type": "integer"
                },
                "example": 1
            }
        }
    },
...
}